Revert "contraseñas arregladas, comprobar todo falta"
This reverts commit bd0793dcca73d9f76166c505595cbe9e28115267.

+++ demo/src/main/java/com/demo/service/GameService.java
@@ -376,14 +376,13 @@ public class GameService {
 		if(puntosRepo.todosVotado(idPartida) && partidaRepo.findById(idPartida) != null) {
 			Puntos p = puntosRepo.getPuntosJugador(idPartida, identificador);
 			List<Integer> resp = new ArrayList<>();
-			//if(puntosRepo.todosConsultado(idPartida)) {
+			if(puntosRepo.todosConsultado(idPartida)) {
 				partidaRepo.deleteRespuestasPartida(idPartida);
 				partidaRepo.deleteJugadoresPartida(idPartida);
 				partidaRepo.deleteHilosPartida(idPartida);
 				partidaRepo.deletePartida(idPartida);
 				puntosRepo.delete(idPartida);
-			//}
-			if(p !=null) {
+			}if(p !=null) {
 				resp.add(p.calcularEstrellas());
 				resp.add(p.calcularMonedas());
 			}
